// friendout-frontend/src/features/subActivity/types/subActivityDetails.type.ts
import type { ParticipationStatus } from "@/features/participant/enum/participationStatus.enum";
import type { Participant } from "@/features/participant/types/Participant.type";
import type { Localisation } from "@/features/localisation/types/localisation.type";

export interface SubActivityDetails {
    id: string;
    name: string;
    localisation?: Localisation | null;
    startTime: string;
    endTime: string;
    description: string;
    price: number;
    activityId: string;
    participants: Array<Participant>;
    UserParticipationStatus?: ParticipationStatus;
}